+++ src/OpenTelemetry.Exporter.Prometheus.AspNetCore/PrometheusExporterMiddleware.cs
@@ -3,7 +3,9 @@
 
 using System.Diagnostics;
 using System.Globalization;
+using System.IO.Compression;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Http.Headers;
 using Microsoft.Net.Http.Headers;
 using OpenTelemetry.Exporter.Prometheus;
 using OpenTelemetry.Internal;
@@ -71,8 +73,10 @@ internal sealed class PrometheusExporterMiddleware
 
             using var linkedCts = CancellationTokenSource.CreateLinkedTokenSource(requestCancelled.Token, httpContext.RequestAborted);
 
-            var openMetricsRequested = AcceptsOpenMetrics(httpContext.Request);
-            var collectionResponse = await this.exporter.CollectionManager.EnterCollect(openMetricsRequested).ConfigureAwait(false);
+            var requestHeaders = httpContext.Request.GetTypedHeaders();
+
+            var openMetricsRequested = AcceptsOpenMetrics(requestHeaders);
+            var collectionResponse = await this.exporter.CollectionManager.EnterCollect(openMetricsRequested);
 
             try
             {
@@ -90,7 +94,7 @@ internal sealed class PrometheusExporterMiddleware
                         ? OpenMetricsContentType
                         : "text/plain; charset=utf-8; version=0.0.4";
 
-                    await response.Body.WriteAsync(dataView.Array.AsMemory(0, dataView.Count), linkedCts.Token).ConfigureAwait(false);
+                    await WriteResponseAsync(response, dataView.Array.AsMemory(0, dataView.Count), AcceptsGZip(requestHeaders), linkedCts.Token);
                 }
                 else
                 {
@@ -125,9 +129,9 @@ internal sealed class PrometheusExporterMiddleware
         }
     }
 
-    internal static bool AcceptsOpenMetrics(HttpRequest request)
+    internal static bool AcceptsOpenMetrics(RequestHeaders headers)
     {
-        var acceptHeader = request.GetTypedHeaders().Accept;
+        var acceptHeader = headers.Accept;
 
         if (acceptHeader is not { Count: > 0 })
         {
@@ -185,5 +189,45 @@ internal sealed class PrometheusExporterMiddleware
         }
 
         return hasSupportedOpenMetricsVersion && hasSupportedOpenMetricsEscaping;
+    }
+
+    private static bool AcceptsGZip(RequestHeaders headers)
+    {
+        if (headers.AcceptEncoding is { Count: > 0 } acceptEncoding)
+        {
+            foreach (var parameter in acceptEncoding)
+            {
+                if (parameter.Value.Equals("gzip", StringComparison.OrdinalIgnoreCase))
+                {
+                    return true;
+                }
+            }
+        }
+
+        return false;
+    }
+
+    private static async Task WriteResponseAsync(
+        HttpResponse response,
+        ReadOnlyMemory<byte> content,
+        bool compress,
+        CancellationToken cancellationToken)
+    {
+        if (compress)
+        {
+            response.Headers.Append("Content-Encoding", "gzip");
+
+            await using var gzip = new GZipStream(
+                response.Body,
+                CompressionLevel.Optimal,
+                leaveOpen: true);
+
+            await gzip.WriteAsync(content, cancellationToken);
+            await gzip.FlushAsync(cancellationToken);
+        }
+        else
+        {
+            await response.BodyWriter.WriteAsync(content, cancellationToken);
+        }
     }
 }