+++ test/OpenTelemetry.Exporter.Prometheus.AspNetCore.Tests/PrometheusIntegrationTests.cs
@@ -3,6 +3,8 @@
 
 using System.Diagnostics;
 using System.Diagnostics.Metrics;
+using System.IO.Compression;
+using System.Net;
 using System.Net.Http.Json;
 using System.Text.Json;
 using Microsoft.AspNetCore.Builder;
@@ -10,6 +12,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Hosting.Server;
 using Microsoft.AspNetCore.Hosting.Server.Features;
 using Microsoft.AspNetCore.WebUtilities;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using OpenTelemetry.Exporter.Prometheus.Tests;
 using OpenTelemetry.Metrics;
@@ -23,13 +26,62 @@ namespace OpenTelemetry.Exporter.Prometheus.AspNetCore.Tests;
 [Collection(PromToolCollection.Name)]
 public class PrometheusIntegrationTests(PromToolFixture promtool, ITestOutputHelper outputHelper)
 {
+    [Fact]
+    public async Task Scrape_Endpoint_Uses_GZip_When_Requested()
+    {
+        // Arrange
+        var builder = WebApplication.CreateBuilder();
+
+        // Listen on any available port
+        builder.WebHost.UseUrls("http://127.0.0.1:0");
+
+        builder.Services
+            .AddOpenTelemetry()
+            .WithMetrics((builder) => builder.AddPrometheusExporter());
+
+        using var app = builder.Build();
+
+        app.MapPrometheusScrapingEndpoint();
+
+        await app.StartAsync();
+
+        var server = app.Services.GetRequiredService<IServer>();
+        var addresses = server.Features.Get<IServerAddressesFeature>();
+
+        var baseAddress = addresses!.Addresses
+            .Select((p) => new Uri(p))
+            .Last();
+
+        using var httpClient = new HttpClient();
+
+        httpClient.DefaultRequestHeaders.Add("Accept-Encoding", "gzip");
+
+        using var response = await httpClient.GetAsync(new Uri(baseAddress, "metrics"));
+
+        // Assert
+        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
+        Assert.NotNull(response.Content);
+        Assert.NotNull(response.Content.Headers.ContentEncoding);
+        Assert.Equal<string>(["gzip"], response.Content.Headers.ContentEncoding);
+        Assert.NotNull(response.Content.Headers.ContentType);
+        Assert.Equal("text/plain; charset=utf-8; version=0.0.4", response.Content.Headers.ContentType.ToString());
+
+        using var compressed = await response.Content.ReadAsStreamAsync();
+        using var decompressed = new GZipStream(compressed, CompressionMode.Decompress);
+        using var reader = new StreamReader(decompressed);
+
+        var content = await reader.ReadToEndAsync();
+
+        Assert.NotEmpty(content);
+        Assert.EndsWith(content, "# EOF\n", StringComparison.Ordinal);
+    }
+
     [EnabledOnDockerPlatformTheory(DockerPlatform.Linux)]
     [InlineData("")]
     [InlineData("OpenMetricsText0.0.1")]
     [InlineData("OpenMetricsText1.0.0")]
     [InlineData("PrometheusText0.0.4")]
     [InlineData("PrometheusText1.0.0")]
-
     public async Task Prometheus_Can_Scrape_Metrics(string scrapeProtocol) => await GenerateMetricsAsync(async (baseAddress) =>
     {
         // Arrange
@@ -219,7 +271,6 @@ public class PrometheusIntegrationTests(PromToolFixture promtool, ITestOutputHel
     [InlineData("application/openmetrics-text;version=0.0.4")]
     [InlineData("application/openmetrics-text;version=1.0.0", Skip = "https://github.com/prometheus/prometheus/issues/8932")]
     [InlineData("application/openmetrics-text;version=1.0.0;escaping=allow-utf-8;q=0.5,application/openmetrics-text;version=0.0.1;q=0.4,text/plain;version=1.0.0;escaping=allow-utf-8;q=0.3,text/plain;version=0.0.4;q=0.2,/;q=0.1", Skip = "https://github.com/prometheus/prometheus/issues/8932")]
-
     public async Task Promtool_Considers_Scrape_Response_Valid(string accept) => await GenerateMetricsAsync(async (baseAddress) =>
     {
         // Act